+++ apps/api/src/health/entities.py
@@ -1,6 +1,21 @@
+from dataclasses import dataclass
+
 from peewee import BigIntegerField, CharField, DecimalField
 
 from src.core.entities import Entity
+
+
+@dataclass(slots=True)
+class DiskUtilizationSummary:
+    stale: bool
+    severity: str | None
+    filesystem: str | None
+    mountpoint: str | None
+    total_bytes: int | None
+    used_bytes: int | None
+    available_bytes: int | None
+    used_percent: float | None
+    last_received_at: int | None
 
 
 class DiskUtilization(Entity):
